+++ modules/apps/29-fee/keeper/keeper.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/cosmos/ibc-go/v6/modules/apps/29-fee/types"
 	channeltypes "github.com/cosmos/ibc-go/v6/modules/core/04-channel/types"
+	porttypes "github.com/cosmos/ibc-go/v6/modules/core/05-port/types"
 	host "github.com/cosmos/ibc-go/v6/modules/core/24-host"
 )
 
@@ -26,7 +27,7 @@ type Keeper struct {
 	cdc      codec.BinaryCodec
 
 	authKeeper    types.AccountKeeper
-	ics4Wrapper   types.ICS4Wrapper
+	ics4Wrapper   porttypes.ICS4Wrapper
 	channelKeeper types.ChannelKeeper
 	portKeeper    types.PortKeeper
 	bankKeeper    types.BankKeeper
@@ -34,8 +35,14 @@ type Keeper struct {
 
 // NewKeeper creates a new 29-fee Keeper instance
 func NewKeeper(
+<<<<<<< HEAD
 	cdc codec.BinaryCodec, key storetypes.StoreKey, paramSpace paramtypes.Subspace,
 	ics4Wrapper types.ICS4Wrapper, channelKeeper types.ChannelKeeper, portKeeper types.PortKeeper, authKeeper types.AccountKeeper, bankKeeper types.BankKeeper,
+=======
+	cdc codec.BinaryCodec, key storetypes.StoreKey,
+	ics4Wrapper porttypes.ICS4Wrapper, channelKeeper types.ChannelKeeper,
+	portKeeper types.PortKeeper, authKeeper types.AccountKeeper, bankKeeper types.BankKeeper,
+>>>>>>> 88525d2 (Simplify SendPacket API (#1703))
 ) Keeper {
 	return Keeper{
 		cdc:           cdc,